VSR: Fix correctness bug (latest normal view)
The latest normal view (as specified in the 2012 paper) is different to
the view number contained in the prepare headers that we include in the
body of the do_view_change message.

The former shows how recent a view change the replica participated in,
which may be much higher than the view number in the prepare header,
which is what we were formerly using to decide the latest head op
after a view change.