VSR: Scan the whole pipeline when repairing a prepare
Avoid any potential for future off-by-one errors if the pipeline is
ever shifted slightly across view changes.

There should be no impact on performance when a replica is repairing
a large portion of its log, since we call this for uncommitted ops,
and also have fast paths to eliminate the case where the op is
before or after the whole pipeline.